move keyboard controller to new world
