+++ subiquity/client/controllers/keyboard.py
@@ -13,56 +13,31 @@
 # You should have received a copy of the GNU Affero General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
+
 import logging
 
-import attr
-
-from subiquitycore.context import with_context
-
-from subiquity.common.keyboard import set_keyboard
-from subiquity.controller import SubiquityTuiController
-from subiquity.keyboard import KeyboardList
-from subiquity.models.keyboard import KeyboardSetting
+from subiquity.client.controller import SubiquityTuiController
+from subiquity.client.keyboard import KeyboardList
+from subiquity.common.types import KeyboardSetting
+from subiquity.common.keyboard import (
+    set_keyboard,
+    )
 from subiquity.ui.views import KeyboardView
 
-log = logging.getLogger('subiquity.controllers.keyboard')
+log = logging.getLogger('subiquity.client.controllers.keyboard')
 
 
 class KeyboardController(SubiquityTuiController):
 
-    autoinstall_key = model_name = "keyboard"
-    autoinstall_schema = {
-        'type': 'object',
-        'properties': {
-            'layout': {'type': 'string'},
-            'variant': {'type': 'string'},
-            'toggle': {'type': ['string', 'null']},
-            },
-        'required': ['layout'],
-        'additionalProperties': False,
-        }
+    endpoint_name = 'keyboard'
+
     signals = [
         ('l10n:language-selected', 'language_selected'),
         ]
 
     def __init__(self, app):
-        self.needs_set_keyboard = False
         super().__init__(app)
         self.keyboard_list = KeyboardList()
-
-    def load_autoinstall_data(self, data):
-        if data is None:
-            return
-        setting = KeyboardSetting(**data)
-        if self.model.setting != setting:
-            self.needs_set_keyboard = True
-        self.model.setting = setting
-
-    @with_context()
-    async def apply_autoinstall_config(self, context):
-        if self.needs_set_keyboard:
-            await set_keyboard(
-                self.app.root, self.model.setting, self.opts.dry_run)
 
     def language_selected(self, code):
         log.debug("language_selected %s", code)
@@ -73,10 +48,12 @@ class KeyboardController(SubiquityTuiController):
         log.debug("loading language %s", code)
         self.keyboard_list.load_language(code)
 
-    def make_ui(self):
+    async def make_ui(self):
         if self.keyboard_list.current_lang is None:
             self.keyboard_list.load_language('C')
-        return KeyboardView(self, self.model.setting)
+        initial_setting = await self.endpoint.GET()
+        view = KeyboardView(self, initial_setting)
+        return view
 
     def run_answers(self):
         if 'layout' in self.answers:
@@ -93,12 +70,7 @@ class KeyboardController(SubiquityTuiController):
         if apply:
             self.app.aio_loop.create_task(self.set_keyboard(setting))
         else:
-            self.model.setting = setting
-            self.configured()
-            self.app.next_screen()
+            self.app.next_screen(self.endpoint.POST(setting))
 
     def cancel(self):
         self.app.prev_screen()
-
-    def make_autoinstall(self):
-        return attr.asdict(self.model.setting)
