+++ subiquity/controllers/__init__.py
@@ -16,7 +16,6 @@
 from ..controller import RepeatedController
 from .filesystem import FilesystemController
 from .identity import IdentityController
-from .keyboard import KeyboardController
 from .mirror import MirrorController
 from .network import NetworkController
 from .proxy import ProxyController
@@ -28,7 +27,6 @@ from .zdev import ZdevController
 __all__ = [
     'FilesystemController',
     'IdentityController',
-    'KeyboardController',
     'MirrorController',
     'NetworkController',
     'ProxyController',
